chore(cli): Phase 2 — reverse public-API dependency direction
src/plugins.ts is the package's public surface (re-exported from
src/index.ts). It previously reached into src/runtime/ for TableScope
and PluginRuntime — meaning the public types depended on private
internals.

Move the TableScope, LoadedPlugin, and PluginRuntime interface
declarations into plugins.ts so the public surface owns them. The
runtime modules now import these types from ../plugins.js instead of
defining them. runtime/table-scope.ts re-exports TableScope so
existing internal consumers keep working unchanged.
