+++ packages/cli/src/runtime/plugin-runtime.ts
@@ -8,103 +8,26 @@ import {
 	type ChxLegacyPluginRegistration,
 	type ChxPluginRegistration,
 	canonicalizeDefinitions,
-	type MigrationPlan,
 	type ResolvedChxConfig,
-	type SchemaDefinition,
 } from '@chkit/core'
 
 import type {
-	ChxGetContextInput,
-	ChxOnAfterApplyContext,
-	ChxOnBeforeApplyContext,
 	ChxOnBeforePluginCommandContext,
 	ChxOnBeforePluginCommandResult,
-	ChxOnCheckContext,
 	ChxOnCheckResult,
-	ChxOnCompleteContext,
-	ChxOnConfigLoadedContext,
-	ChxOnInitContext,
-	ChxOnPlanCreatedContext,
-	ChxOnSchemaLoadedContext,
 	ChxPlugin,
-	ChxPluginCommand,
-	ChxPluginCommandContext,
+	LoadedPlugin,
 	PluginContext,
+	PluginRuntime,
+	TableScope,
 } from '../plugins.js'
 import { isInlinePluginRegistration } from '../plugins.js'
 import { debug, isDebugEnabled } from './debug.js'
-import type { TableScope } from './table-scope.js'
-
-interface LoadedPlugin {
-	options: Record<string, unknown>
-	plugin: ChxPlugin
-}
 
 const UNFILTERED_TABLE_SCOPE: TableScope = {
 	enabled: false,
 	matchedTables: [],
 	matchCount: 0,
-}
-
-export interface PluginRuntime {
-	plugins: ReadonlyArray<LoadedPlugin>
-	getCommand(
-		pluginName: string,
-		commandName: string,
-	): { command: ChxPluginCommand; plugin: LoadedPlugin } | null
-	resolveContext(
-		input: Omit<ChxGetContextInput, 'defaults'>,
-	): Promise<PluginContext>
-	disposeContext(ctx: PluginContext): Promise<void>
-	runOnInit(context: Omit<ChxOnInitContext, 'options'>): Promise<void>
-	runOnComplete(
-		context: Omit<ChxOnCompleteContext, 'options' | 'exitCode'> & {
-			exitCode?: number
-		},
-	): Promise<void>
-	runOnConfigLoaded(
-		context: Omit<ChxOnConfigLoadedContext, 'options' | 'tableScope'> & {
-			tableScope?: TableScope
-		},
-	): Promise<void>
-	runOnSchemaLoaded(
-		context: ChxOnSchemaLoadedContext,
-	): Promise<SchemaDefinition[]>
-	runOnPlanCreated(
-		context: Omit<ChxOnPlanCreatedContext, 'plan' | 'tableScope'> & {
-			tableScope?: TableScope
-		},
-		plan: MigrationPlan,
-	): Promise<MigrationPlan>
-	runOnBeforeApply(context: ChxOnBeforeApplyContext): Promise<string[]>
-	runOnAfterApply(context: ChxOnAfterApplyContext): Promise<void>
-	runOnCheck(
-		context: Omit<ChxOnCheckContext, 'options' | 'tableScope'> & {
-			tableScope?: TableScope
-		},
-	): Promise<ChxOnCheckResult[]>
-	runOnCheckReport(
-		results: ChxOnCheckResult[],
-		print: (line: string) => void,
-	): Promise<void>
-	runOnBeforePluginCommand(
-		pluginName: string,
-		commandName: string,
-		context: Omit<
-			ChxOnBeforePluginCommandContext,
-			'targetPlugin' | 'command' | 'options'
-		>,
-	): Promise<ChxOnBeforePluginCommandResult>
-	runPluginCommand(
-		pluginName: string,
-		commandName: string,
-		context: Omit<
-			ChxPluginCommandContext,
-			'pluginName' | 'options' | 'tableScope'
-		> & {
-			tableScope?: TableScope
-		},
-	): Promise<number>
 }
 
 function parseCliMajor(version: string): number {
